Refactor Codes
