+++ src/components/Admin/VerifySensei/NewSenseiTable.js
@@ -47,7 +47,7 @@ const NewMentorTable = () => {
         <Column title="Email" dataIndex="email" key="email" />
         <Column title="Created At" dataIndex="createdAt" key="createdAt" />
         <Column title="Admin Verified" dataIndex="adminVerified" key="adminVerified" />
-        <Column title="Privacy" dataIndex="privacy" key="privacy" />
+        <Column title="Chat Privacy" dataIndex="chatPrivacy" key="chatPrivacy" />
         <Column title="Status" dataIndex="status" key="status" />
         <Column
           title="Details"
