+++ src/pages/admin/verify-sensei/sensei-page.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import MentorProfile from '../../../components/Admin/VerifyMentors/MentorProfile'
+import MentorProfile from '../../../components/Admin/VerifySensei/SenseiProfile'
 
 const MentorPage = () => {
   return (
